Moved lots of stuff into a new shortcuts module
